requirements.txt
discord.py
configparser
requests
python-dotenv
dill
unidecode
